Most, perhaps all, of the way towards sorting out the issue with include paths when compiling UINET sources that shim between the host OS and the FreeBSD kernel sources in the UINET tree.
In order to be able to deterministically include the desired host or
UINET kernel include files in a shim source, the build now creates a
copy of the UINET kernel include tree (including override headers) and
uses directory renaming within that tree to put the UINET kernel
includes in a unique namespace (this is known as the shim include
tree).  The shim sources then #include from that unique namespace when
including UINET kernel headers.

The UINET API includes have been moved to a separate api_include
directory.

With the above facility in place, UINET kernel sources are compiled
against the UINET kernel header trees, with the compiler's standard
include paths disabled.  UINET shim sources are compiled against the
shim include tree, with the compiler's standard include paths enabled.
Users of libuinet are compiled against UINET's api_include tree, with
the compiler's standard include paths enabled, and without reference
to the shim include tree.