Added passive-to-active transition support, including transitioning on syncache timeout with a synfilter-configurable timeout.
Passive socket pairs now hold a reference to each other to that the
state of each is available for use during transition to active, even
if flow in one direction has been shut down by one of the remotes.

It is now possible for a non-passive listen socket to spawn passive
sockets and passive listen sockets to spawn non-passive socket - that
is, a syn filter can now decide whether to treat a given connection as
passive or not.  If no syn filter is used, new sockets spawned by a
listen socket still inherit their passive mode setting from the listen
socket.