Unbreak synfilter decision values in libuinet.
The kernel synfilter decision values (SYNF_*) and the libuinet API
versions of the same (UINET_SYNF_*) must match, as they are not
translated at runtime.  They went out of sync with the initial
version of passive receive functionality, breaking everything other
than SYNF_ACCEPT.

Assertions at compile and/or runtime, or translation at runtime to
maintain a stable libuinet ABI, may be desirable in the future.