inp->inp_socket may be NULL here, check for that.
The code here assumes it can dereference inp->inp_socket without a
check for inp->inp_socket being NULL, even though later code admits
this is a potentiality with which this code must deal.  While this
is needed in practice in the outer conditional, not using passive
mode myself at the moment I am unsure if it is needed in the inner
one, and am unable to determine this empirically, unlike with the
outer one.