pdtest has become multitool, and multitool subsumed echo and passive
The unique parts of the existing demo programs are being given a
common interface and moved to a new library, libuinet_demo.  multitool
provides a front end for creating and arranging event loops, stack
instances, and the demo programs in libuinet_demo and running them.
So far echo and passive (which has been split into passive and
passive-extract) have been migrated.

Some small modifications were made to the make system to support
building libraries and dealing with include and link path
configuration when using them in other programs in the tree.