STS-mode vnets now run with socket-level locks disabled
This includes the global socket lock, the per-socket lock, the sockbuf
I/O interlacing lock, the finer-grained sockbuf mutex, and the accept
lock.

The accept lock and the global socket lock that protects
numopensockets and so_gencnt are now per-vnet locks, and
numopensockets and so_gencnt are now per-vnet as well.

The rules for setting curvnet or not when using the uipc sockets API
have shifted slightly and the code has been adjusted accordingly.