More stack locks disabled if running in STS-mode
There's a fair amount of noise in this commit from renaming
VNET_IS_STS() to CURVNET_IS_STS().

Most if-related locks are now disabled when running in STS-mode.  The
one with the biggest payoff is likely the ifqueue lock.  struct
ifqueue and struct ifaltq have grown a vnet pointer member to support
disabling queue locks in STS mode, which if set allows detection of
when queue locks can be avoided.  That vnet pointer doesn't get set by
all ifqueue users, but that's functionally OK as it just potentially
results in unnecessary locking when running in STS-mode.  The if_snd
in struct ifnet gets fully configured, and that's where most of the
bang for this buck is.  The locks protecting the list of ifnets in
each vnet are still global locks - if_vmove() currently relies on that
and it doesn't seem worth the effort to sort out at this point.

The igmp, in_multi, and in_ifaddr locks have been converted to
per-vnet locks and are disabled when running in STS mode.