The so member in ev_uinet watchers is not initialized until the watcher is started for the first time.  This introduces a hazard for applications, such as the passive-extract demo, which under some circumstances close a socket which may not have any started watchers.
The so member in ev_uinet watchers is a convenenience member that
duplictes the so in the attached ev_uinet_ctx structure.  The attached
ev_uinet_ctx structure does not have a similar lifetime issue - its
contents are valid for the entire lifetime of all associated watchers.
An accessor has been added to retrieve the so pointer from an
ev_uinet_ctx structure, so that given an initialized watcher, its
associated socket may always be closed regardless of whether any
watchers for that socket were ever started.

The so member in ev_uinet watchers is very likely to go away soon.