Ensure pending ACK is sent when skipping TIMEWAIT.
tcp_twstart() may be entered with TF_ACKNOW or TF_DELACK set.  If
timewait is to be skipped (V_nolocaltimewait is set and the foerign IP
is local, the socket is in passive mode, or TF_NO_TIMEOUT is set), and
either type of ACK is pending, an immediate ACK is now sent prior to
cleaning up the socket state.  connection state.

As a side effect, skipping timewait in passive mode is now a bit more
efficient.