Added rx-side pcap capture support
There is now a capability to send received packets to pcap
capture files via pcap packet interface instances.  Pcap packet
interface instances can now be configured to transmit-to-file,
either to a single file, or to individual files based on the flow
serial number stored in the packet descriptor.  All packets
transmitted from the stack, or injected to the tx side of, such
interfaces will be written to the associated pcap file(s).  Pcap
packet interface instances configured for transmit-to-file can be
further configured to perform the file I/O in a separate thread,
primarily in order to keep file I/O performance out of the data
path in STS-mode stack instances.

TCP sockets can now be configured to zero-copy replicate all
received packets that match to the socket to the transmit side of
any packet interface.  In particular, sockets can be configured
to zero-copy queue all received packets until the application
decides to replicate them to the given interface or dispose of
them.  A limit on the total number of bytes replicated by a
socket can be configured on a per-socket basis.  Combined with
the new pcap transmit-to-file support described above, this can
be used to selectively capture the receive-side of individual
flows or portions of flows to pcap files.

Note that currently there is no support for replication of the
transmit side of sockets or packet interfaces - that is awaiting
transmit-side zero-copy support through the stack.

Specific changes:

- Update of pcap packet interface to support batch-mode operation
and transmit-to-file.

- Multitool and libuinet_demo app improvements for exercising the
new pcap and rx-side packet replication support.

- New IP-level socket options and uinet API entry point for
configuring rx-side socket packet replication.

- inpcbs have a new serial number attribute that assigns a
non-repeating, stack-wide-unique serial number to each inpcb.
For passive reassembly sockets, both sockets in the pair will
have the same serial number.

- There is now a global epoch number that can be configured via
uinet_init() and a per-instance index value that is assigned at
instance creation time. These are used, along with the new inpcb
serial number, to generate unique filesnames for pcap capture
files when using file-per-flow transmit-to-file mode on pcap
packet interfaces.

- Receive-side timestamping can now be enabled and configured on
a per-packet-interface basis.  These timestamps are written to
pcap capture files when using pcap packet interfaces in
transmit-to-file mode.  Several timestamp modes are supported -
see UINET_IF_TIMESTAMP_xxx in uinet_api_types.h.

- The packet interface rx batch size parameter has been promoted
from netmap-only to a generic option for all interface types.

- A new packet descriptor data structure (uinet_pd_xlist) has
been added to support chains of fixed-size descriptor lists.

- Packet descriptors have been extended to include a flow serial
number, and packet descriptor contexts have been extended to
include a timestamp and the original packet length (so it can be
retrieved via an mbuf - ideally mbufs would have this field).

- In libuinet_demo, support for base/common arguments has been
added.

- The TCP syncache now retains the syn (and in passive-reassembly
mode, the syn|ack) when copy mode is enabled.

- Added strndup to libkern

- Fixed some compiler complaints and typos