Add auto-flush-to-file for pcap socket captures
The implemented mechanism for communicating a flow flush request is
general and the only current consumer of this information is the pcap
packet interface.  Note that the auto-flush mechanism relies on the
interface injection queue, which can lose the request if injected when
the queue is full.  Note also that passive reassembly socket pairs
will currently generate two flow flush requests for the given flow
when closed - one from each socket.

Most of the diff in the netmap driver is an indentation change.

Along the way:

- Added rx-side packet replication support to the passive and
passive-extract demos.

- Added new parameters to multitool configuration print routine.

- Added support for UINET_IF_TIMESTAMP_HW on the pcap receive side
(preserves the pcap timestamps on received packets).

- Housekeeping on unused network interface config parameters, unneeded
 trace messages, and some whitespace issues.